Assignment 1
The first assignment was divided into 2 parts. The first one was to determine the degree sequence of a graph which is esentially the total number of incoming and outgoing connections for a particular node. So I had to calulate the degree for each node and sort in descending order which becomes the degree sequence of the graph.
The 2nd part was to find all possible nodes from a given node with two hops. The implementation was done for an adjacency List and adjacency Matrix representation for a graph.