Assignment 2
In this assignment I had to design my own class to implement a Breadth First Search (BFS) given a start point and end point. I was given skeleton code for only the Graph class which didn't have even it's instance variables or constructors defined! So to do this assignment I authored 2 classes which is MapNode which holds all properties of individual Nodes and RoadNode which holds properties of a Road. Then using these I implemented the BFS algorithm and then tested it to check if everything was fine. Finally I checked the implementation on my Front end and it worked prefectly fine showing me the route from the start to the end. This assignment helped me learn something very useful which was how to desgin classes and make different objects and classes work together to get the required result