Assignment 3
In this assignment I had to implement dijkstra and A* algorithms and compare thier efficiency. Dijkstra works on a priority queue and hence is better than BFS for finding the shortest path but it explores all the nodes before getting to its goal as the algorithm doesn't consider the distance from the goal. This is overcome by A*, as at a particular node it takes into account both, the distance from start and the distance to reach the goal, hence it visits less nodes and is more efficient